fix(telegram): read module version once
