+++ telegram/client.go
@@ -5,8 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"runtime/debug"
-	"strings"
 	"sync"
 	"time"
 
@@ -127,25 +125,6 @@ type Client struct {
 // API returns *tg.Client for calling raw MTProto methods.
 func (c *Client) API() *tg.Client {
 	return c.tg
-}
-
-// getVersion optimistically gets current client version.
-//
-// Does not handle replace directives.
-func getVersion() string {
-	info, ok := debug.ReadBuildInfo()
-	if !ok {
-		return ""
-	}
-	// Hard-coded package name. Probably we can generate this via parsing
-	// the go.mod file.
-	const pkg = "github.com/gotd/td"
-	for _, d := range info.Deps {
-		if strings.HasPrefix(d.Path, pkg) {
-			return d.Version
-		}
-	}
-	return ""
 }
 
 // Port is default port used by telegram.
